Refactor code to use generators and allow interruptions
This makes the data generation function into a generator, and the
writers now iterate over the values returned by the generator while
handling KeyboarInterrupt exceptions.  This makes it possible to ^C
the process and still get output in the output file.

Unfortunately this means we can't figure out how many columns we need
to write in the spreadsheet, because we don't get to wait for all the
data before we start to write.  Currently I don't see a way around
that, so unfortunately it's a limitation we have to live with for now.